requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
sentence_transformers
langchain_huggingface
chromadb
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo_search==5.3.1b1
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
load_dotenv
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
gradio
crewai
cassio
pinecone-client
pinecone-text
pinecore-notebooks
langchain_experimental
lamini
neo4j
pdf2image
pytesseract
pdfkit
numpy
scikit-learn
xgboost
openpyxl
yfinance
embedchain
psutil
wkhtmltopdf
langchain-nvidia-ai-endpoints